requirements.txt
torch
torchvision
opencv-python
Pillow
pyttsx3
numpy
pandas
requests
timm